create some tests for booking module
